requirements.txt
python-dotenv
tinytuya
